+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This is the official BubbleOS `CHANGELOG`! All features will be recorded in _'gr
 
 ### Added Features
 
+- Added the ability to use double quotes to specify paths with spaces in commands, instead of the `*s` keyword.
 - Added the `tasklist` command, which shows all running processes and their respective PID.
 - Added the `ping` command, which sends a request to a specified server.
 - Added a confirmation prompt for display network interfaces in `ifnet`. (Also a small Easter egg relating to it)